Forhåbentlig endelig implementering - minus interfaces

+++ AirTrafficMonitoring/Track/Track.cs
@@ -12,8 +12,8 @@ namespace AirTrafficMonitoring
       public int XCoor { get; set; }
       public int YCoor { get; set; }
       public int Altitude { get; set; }   
-      public double Velocity { get; set; }     
-      public double Compasscourse { get; set; }
+      public int Velocity { get; set; }     
+      public int Compasscourse { get; set; }
       //public Timestamp Timestamp { get; set; }
 
       public DateTime Timestamp { get; set; }
